combine PathRewriterTrait & CSSTrait
