+++ inc/Engine/Optimization/CacheDynamicResource.php
@@ -4,7 +4,6 @@ namespace WP_Rocket\Engine\Optimization;
 
 use WP_Rocket\Admin\Options_Data;
 use WP_Rocket\Event_Management\Subscriber_Interface;
-use WP_Rocket\Engine\Optimization\PathRewriterTrait;
 
 /**
  * Create a static file for CSS/JS generated by a PHP file
@@ -13,7 +12,6 @@ use WP_Rocket\Engine\Optimization\PathRewriterTrait;
  */
 class CacheDynamicResource extends AbstractOptimization implements Subscriber_Interface {
 	use CSSTrait;
-	use PathRewriterTrait;
 
 	/**
 	 * Plugin options instance.
