+++ inc/Engine/Optimization/Minify/CSS/Combine.php
@@ -5,7 +5,6 @@ use MatthiasMullie\Minify\CSS as MinifyCSS;
 use WP_Rocket\Admin\Options_Data;
 use WP_Rocket\Engine\Optimization\AssetsLocalCache;
 use WP_Rocket\Engine\Optimization\CSSTrait;
-use WP_Rocket\Engine\Optimization\PathRewriterTrait;
 use WP_Rocket\Logger\Logger;
 
 /**
@@ -15,7 +14,6 @@ use WP_Rocket\Logger\Logger;
  */
 class Combine extends AbstractCSSOptimization {
 	use CSSTrait;
-	use PathRewriterTrait;
 
 	/**
 	 * Assets local cache instance
