+++ inc/deprecated/Engine/Optimization/QueryString/Remove.php
@@ -4,7 +4,7 @@ namespace WP_Rocket\Engine\Optimization\QueryString;
 use WP_Rocket\Admin\Options_Data;
 use WP_Rocket\deprecated\DeprecatedClassTrait;
 use WP_Rocket\Engine\Optimization\AbstractOptimization;
-use WP_Rocket\Engine\Optimization\PathRewriterTrait;
+use WP_Rocket\Engine\Optimization\CSSTrait;
 
 /**
  * Remove query string from static resources.
@@ -16,7 +16,7 @@ use WP_Rocket\Engine\Optimization\PathRewriterTrait;
  */
 class Remove extends AbstractOptimization {
 	use DeprecatedClassTrait;
-	use PathRewriterTrait;
+	use CSSTrait;
 
 	/**
 	 * Plugin options instance.
